Get path without accessed-before-initialization errors
If descendants don't specify `$this->path`, then calling `path()` will fail with an error: `Typed property JustSteveKing\Transporter\Request::$path must not be accessed before initialization`

This is one way to resolve the problem.  Another option is to initialize the variable when it's declared:
```php
    protected string $path = '';
```

Not sure which way is best/preferable...or if there's some other even better way to go. 🤓